- called retrive opportunity in ConnectJobHelperClass
