+++ app/src/org/commcare/activities/LoginActivity.java
@@ -503,7 +503,7 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
     private void setResultAndFinish(boolean navigateToConnectJobs) {
         if (personalIdManager.isloggedIn() && !navigateToConnectJobs && !ConnectAppUtils.INSTANCE.hasConnectJobs(
                 this)) {
-            ConnectAppUtils.INSTANCE.retrieveConnectOpportunities(this);
+            ConnectJobHelper.INSTANCE.retrieveConnectOpportunities(this);
         }
         Intent i = new Intent();
         i.putExtra(REDIRECT_TO_CONNECT_OPPORTUNITY_INFO, navigateToConnectJobs);
