+++ app/src/org/commcare/connect/ConnectJobHelper.kt
@@ -1,5 +1,6 @@
 package org.commcare.connect
 
+import android.app.Activity
 import android.content.Context
 import android.widget.Toast
 import org.commcare.android.database.connect.models.ConnectJobAssessmentRecord
@@ -9,6 +10,7 @@ import org.commcare.android.database.connect.models.ConnectJobPaymentRecord
 import org.commcare.android.database.connect.models.ConnectJobRecord
 import org.commcare.connect.database.ConnectJobUtils
 import org.commcare.connect.database.ConnectUserDatabaseUtil
+import org.commcare.connect.database.JobStoreManager
 import org.commcare.connect.network.ApiConnect
 import org.commcare.connect.network.ConnectNetworkHelper
 import org.commcare.connect.network.IApiCallback
@@ -17,6 +19,7 @@ import org.commcare.google.services.analytics.FirebaseAnalyticsUtil
 import org.javarosa.core.io.StreamsUtil
 import org.javarosa.core.model.utils.DateUtils
 import org.javarosa.core.services.Logger
+import org.json.JSONArray
 import org.json.JSONException
 import org.json.JSONObject
 import java.io.IOException
@@ -336,5 +339,65 @@ object ConnectJobHelper {
                     listener.connectActivityComplete(false)
                 }
             })
+    }
+
+
+    /**
+     * Retrieves Connect job opportunities for the current user and stores them locally.
+     * Requires valid activity to fetch details.
+     */
+    fun retrieveConnectOpportunities(activity: Activity) {
+        val user = ConnectUserDatabaseUtil.getUser(activity)
+        ApiConnect.getConnectOpportunities(activity, user, object : IApiCallback {
+            override fun processSuccess(responseCode: Int, responseData: InputStream?) {
+                try {
+                    val responseAsString = responseData?.let { input ->
+                        String(StreamsUtil.inputStreamToByteArray(input))
+                    } ?: return
+
+                    if (responseAsString.isNotEmpty()) {
+                        val json = JSONArray(responseAsString)
+                        val jobs = mutableListOf<ConnectJobRecord>()
+
+                        for (i in 0 until json.length()) {
+                            try {
+                                val obj = json.getJSONObject(i)
+                                val job = ConnectJobRecord.fromJson(obj)
+                                jobs.add(job)
+                            } catch (e: JSONException) {
+                                Logger.exception("Parsing return from Opportunities request", e)
+                            }
+                        }
+
+                        JobStoreManager(activity).storeJobs(activity, jobs, true)
+                    }
+
+                } catch (e: JSONException) {
+                    throw RuntimeException(e)
+                } catch (e: IOException) {
+                    Logger.exception("Parsing return from Opportunities request", e)
+                }
+            }
+
+            override fun processFailure(responseCode: Int, errorResponse: InputStream?) {
+                Logger.log("ERROR", "Opportunities call failed: $responseCode")
+            }
+
+            override fun processNetworkFailure() {
+                ConnectNetworkHelper.showNetworkError(activity)
+            }
+
+            override fun processTokenUnavailableError() {
+                ConnectNetworkHelper.handleTokenUnavailableException(activity)
+            }
+
+            override fun processTokenRequestDeniedError() {
+                ConnectNetworkHelper.handleTokenDeniedException()
+            }
+
+            override fun processOldApiError() {
+                ConnectNetworkHelper.showOutdatedApiError(activity)
+            }
+        })
     }
 }
